+++ core/src/main/java/tech/ydb/core/auth/StaticCredentials.java
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory;
 
 import tech.ydb.auth.AuthRpcProvider;
 import tech.ydb.auth.YdbAuth;
-import tech.ydb.core.grpc.impl.GrpcAuthRpc;
+import tech.ydb.core.impl.auth.GrpcAuthRpc;
 
 /**
  *
