+++ core/src/main/java/tech/ydb/core/impl/auth/GrpcAuthRpc.java
@@ -1,8 +1,12 @@
-package tech.ydb.core.grpc.impl;
+package tech.ydb.core.impl.auth;
 
 import java.util.concurrent.ExecutorService;
 
 import tech.ydb.core.grpc.GrpcTransport;
+import tech.ydb.core.impl.AnonimousTransport;
+import tech.ydb.core.impl.BaseGrpcTrasnsport;
+import tech.ydb.core.impl.pool.EndpointRecord;
+import tech.ydb.core.impl.pool.ManagedChannelFactory;
 
 /**
  *
