+++ core/src/main/java/tech/ydb/core/impl/pool/EndpointRecord.java
@@ -1,4 +1,4 @@
-package tech.ydb.core.grpc.impl;
+package tech.ydb.core.impl.pool;
 
 import java.util.Objects;
