+++ core/src/main/java/tech/ydb/core/impl/pool/ManagedChannelFactory.java
@@ -1,4 +1,4 @@
-package tech.ydb.core.grpc.impl;
+package tech.ydb.core.impl.pool;
 
 import java.io.ByteArrayInputStream;
 import java.util.function.Consumer;
